Reorganize copy button into separate component and relocate contrast check file
