+++ src/Components/Color/Color.jsx
@@ -1,37 +1,20 @@
-import { useState, useEffect } from "react";
+import { useState } from "react";
 import "./Color.css";
 import ColorForm from "../ColorForm/ColorForm";
-import ContrastCheck from "../ContrastCheck";
+import ContrastCheck from "./ContrastCheck";
+import CopyButton from "./CopyButton";
 
-export default function Color({ color, onDeleteColor, onEditColor, currentThemeId }) {
+export default function Color({
+  color,
+  onDeleteColor,
+  onEditColor,
+  currentThemeId,
+}) {
   const [deleteMode, setDeleteMode] = useState(false);
   const [editMode, setEditMode] = useState(false);
-  const [copied, setCopied] = useState(false);
-
-  useEffect(() => {
-    let timeoutId;
-    if (copied) {
-      timeoutId = setTimeout(() => {
-        setCopied(false);
-      }, 3000);
-    }
-
-    return () => {
-      clearTimeout(timeoutId);
-    };
-  }, [copied]);
 
   function handleExitEditMode() {
     setEditMode(false);
-  }
-
-  async function handleCopy() {
-    try {
-      await navigator.clipboard.writeText(color.hex);
-      setCopied(true);
-    } catch (error) {
-      console.error(error.message);
-    }
   }
 
   return (
@@ -43,9 +26,7 @@ export default function Color({ color, onDeleteColor, onEditColor, currentThemeI
       }}
     >
       <h3 className="color-card-headline">{color.hex}</h3>
-      <button type="button" className="color_copy" onClick={handleCopy}>
-        {copied ? "Copied to clipboard!" : "Copy"}
-      </button>
+      <CopyButton color={color} />
       <h4>{color.role}</h4>
       <p>contrast: {color.contrastText}</p>
       <ContrastCheck colors={[color.hex, color.contrastText]} />
@@ -68,7 +49,9 @@ export default function Color({ color, onDeleteColor, onEditColor, currentThemeI
             type="button"
             className="color_delete"
             onClick={() =>
-              deleteMode ? onDeleteColor(color.id, currentThemeId) : setDeleteMode(true)
+              deleteMode
+                ? onDeleteColor(color.id, currentThemeId)
+                : setDeleteMode(true)
             }
           >
             Delete
